Added helper tool to record image topic to a video
It relies on the standard OpenCV VideoWriter class.

It is configured to encode the video as MJPG by default in an AVI
container (AVI is mandated by OpenCV).

Alternative encoders can be specified with the 'codec' parameter.

The video size is acquired from the topic, but the frame-rate has to be
manually set with the node parameter 'fps' (defaults to 15 fps).