Implement semi-global block matching
Create the necessary code to produce disparity images using the
semi-global BM from OpenCV. To do so I have replicated lot of code in a
sg_disparity nodelet. Obviously it can be made a lot cleaner and it
reamains TODO. Although the point cloud is generated, it is awfully slow
(not in the case of disparity images, which are incredibly fast).
Dynamic parameter tunning is also pending.