Changes and additions from Zach LaCelle at The MITRE Corporation to add NVIDIA VisionWorks support for stereo algorithm acceleration in ROS stereo_image_proc.
This includes:
* Adding a library for visionworks functionality, linked against NVIDIA libovx and libnvx
* Adding supporting calls in processor.h/processor.cpp to switch to this functionality
* Editing dynamic-reconfigure options
* Adding supporting documentation

This software should auto-detect the presence of installed VisionWorks code and, if detected, compile and link against said code. If the requirements are not met, then stereo_image_proc will not link in the new code and will continue to run in the traditional stereo_image_proc method.