Port stereo_image_proc to ROS 2 (#486)
* Rename source files

More descriptive names that will match the classes contained inside.
Also moved all source files associated with the library into a directory of the same name.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Get stereo_image_proc compiling with ROS 2

Made some minimum changes required for compilation. Not feature-complete yet.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Add launch test for disparity node

The test simply checks if the disparity node publishes to the expected topic when it received input.
Made some fixes added to CMakeLists.txt to get node to run probably and pass the test.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Add launch test for PointCloudNode

Similar to the test for the disparity node.
Fix bug in PointCloudNode.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Remove unused files

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Add Python launch file

Replaces the old ROS 1 XML launch file.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Remove standalone executable

This form of manual composition seems redundant to me and can be done with a launch file.
The only thing it appeared to add was the 'advertisement checker', which was not ported.

If there is strong objection to removing this executable, we can add it back.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Add linter tests and lint

There are a few cases where we discard the const-qualifier for the image data.
This happens because OpenCV won't accend const pointers to data.
Rather than changing the API, I've introduced a const_cast instead.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Remove OpenCV 2 support

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Expand macros

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Enable uncrustify test

Fix lint error.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Remove dead code and add TODO

The code removed was never being used.
Added a TODO regarding the use of an unset variable. This has been here since the ROS 1 implementation,
I'm not sure what the logic is intended to do so I've left it for now.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Update package.xml version

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Minor refactor

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Fix typo

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Minor: update include syntax and order for consistency

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Fix rosdep key typo

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Move inline implementation into class

Save on some vertical space.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Avoid extra computation if there are no matching subscriptions

Remove commented code related to subscriber status since this may look different in ROS 2.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Refactor tests for compatbility with Dashing

Now the launch tests should work for both Dashing and Eloquent.
I left some TODO's where we can simplify once we diverge from Dashing.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>

* Use default context

Signed-off-by: Jacob Perron <jacob@openrobotics.org>