Adding Nvidia Image Proc link (#954)
There is an nvidia image pipeline implementation that is GPU optimized
for the Jetsons. I think its worth mentioning to users that are using
the Jetson platforms that there are a few related modules they should
check out if interested.

The Nvidia version is not feature-complete, since some are not
meaningfully optimized by the GPU, but contains a few useful ones.